sync state with local storage, fix message types for signalr, heartbeat and message notifications working,

+++ src/GameManager.Server/Controllers/PlayersController.cs
@@ -1,23 +1,33 @@
 using AutoMapper;
 using GameManager.Server.Data;
 using GameManager.Server.DTO;
+using GameManager.Server.Messages;
 using GameManager.Server.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.SignalR;
 
 namespace GameManager.Server.Controllers;
 
 [Route("api/[controller]")]
 [ApiController]
+[Authorize]
 public class PlayersController : ControllerBase
 {
     private readonly PlayerRepository _playerRepository;
 
     private readonly IMapper _mapper;
 
-    public PlayersController(PlayerRepository playerRepository, IMapper mapper)
+    private readonly IHubContext<GameHub> _hubContext;
+
+    public PlayersController(
+        PlayerRepository playerRepository,
+        IMapper mapper,
+        IHubContext<GameHub> hubContext)
     {
         _playerRepository = playerRepository;
         _mapper = mapper;
+        _hubContext = hubContext;
     }
 
     [HttpGet("{id}")]
@@ -59,6 +69,16 @@ public class PlayersController : ControllerBase
         }
         
         dto = _mapper.Map<PlayerDTO>(player);
+        
+        // Notify other players
+        var message = new PlayerStateChangedMessage()
+        {
+            GameId = player.GameId,
+            Player = dto
+        };
+
+        await _hubContext.Clients.Group(player.GameId.ToString())
+            .SendAsync(nameof(IGameHubClient.PlayerStateChanged), message);
 
         return Ok(dto);
     }
@@ -66,7 +86,24 @@ public class PlayersController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeletePlayer([FromRoute] Guid id)
     {
+        var player = await _playerRepository.GetPlayerById(id);
+
+        if (player == null)
+        {
+            return NotFound();
+        }
+        
         await _playerRepository.RemovePlayer(id);
+        
+        // Notify other players
+        var message = new PlayerLeftMessage()
+        {
+            GameId = player.GameId,
+            PlayerId = player.Id
+        };
+
+        await _hubContext.Clients.Group(player.GameId.ToString())
+            .SendAsync(nameof(IGameHubClient.PlayerLeft), message);
 
         return NoContent();
     }
